move test helpers into their own file
