+++ EcoSoapBankTests/PickupTests.swift
@@ -199,39 +199,6 @@ extension PickupTests {
     }
 }
 
-protocol HostingController {}
-
-extension UIHostingController: HostingController {}
-
-extension XCTestCase {
-    func wait(for expectation: XCTestExpectation, timeout: TimeInterval = 5) {
-        wait(for: [expectation], timeout: timeout)
-    }
-}
-
-
-protocol KeyPathListable {
-    var allKeyPaths: [PartialKeyPath<Self>] { get }
-}
-
-extension KeyPathListable {
-    var allKeyPaths: [PartialKeyPath<Self>] {
-        let mirror = Mirror(reflecting: self)
-
-        return mirror.children
-            .reduce(into: [PartialKeyPath<Self>]()) { keyPaths, labelValuePair in
-                guard let key = labelValuePair.label else { return }
-                keyPaths.append(\Self.[checkedMirrorDescendent: key] as PartialKeyPath)
-        }
-    }
-
-    private subscript(checkedMirrorDescendent key: String) -> Any {
-        Mirror(reflecting: self).descendant(key)!
-    }
-}
-
-extension Property: KeyPathListable {}
-
 // MARK: - Mock Delegate
 
 class MockScheduleVMDelegate: SchedulePickupViewModelDelegate {
